// frontend/src/shared/lib/api/model/provenanceCitationResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * prot-cellar
 * OpenAPI spec version: 0.1.0
 */
import type { ProvenanceCitationResponsePmid } from './provenanceCitationResponsePmid';
import type { ProvenanceCitationResponseDoi } from './provenanceCitationResponseDoi';
import type { ProvenanceCitationResponseUrl } from './provenanceCitationResponseUrl';
import type { ProvenanceCitationResponseLabel } from './provenanceCitationResponseLabel';

export interface ProvenanceCitationResponse {
  pmid?: ProvenanceCitationResponsePmid;
  doi?: ProvenanceCitationResponseDoi;
  url?: ProvenanceCitationResponseUrl;
  label?: ProvenanceCitationResponseLabel;
}